tests updated

+++ app/Controller/Blog.php
@@ -7,29 +7,49 @@ use Core\AbstractController;
 use Core\DataBase;
 use Core\Normalizer;
 use Core\Validator;
+use PharIo\Manifest\ElementCollection;
 
 class Blog extends AbstractController
 {
-    public array $errors = [];
+    private array $errors = [];
+    private const PATH_TO_IMG = '../../html/img/';
+    private string $pathToPatterns = PATH_TO_ERROR_PATTERNS . DIRECTORY_SEPARATOR . "Blog";
 
-     /**
+    /**
+     * @return mixed[]
+     */
+    public function getErrors(): array
+    {
+        return $this->errors;
+    }
+
+    /**
      * В случае, если пользователь не авторизован, перенаправит на страницу авторизации,
      * в противном случае запускается контроллер вывода соообщений
      * @throws \Core\RedirectException
      */
-    public function indexAction(): void
+    public function indexAction(bool $print = true)
     {
+
+        if(!$print)
+        {
+            ob_start();
+        }
 
         if (!isset($_SESSION['id'])) {
             $this->errors[] = USER_NOT_AUTHORIZED;
-            if ($this->getView() !== null) {
-                echo $this->getView()->render('Blog/blogErr.phtml', ['errCodes' => $this->errors]);
-            }
 
-            die;
+            echo $this->getView()->render($this->pathToPatterns,
+                [
+                    'errCodes' => \Core\Validator::getErrors($this->errors)
+                ],
+                'blogErr.twig', 0, 0, 0);
+
+            return null;
         }
 
-        $this->redirect('/html/blog/getAll');
+        return $this->redirect('/html/blog/getAll');
+
     }
 
 
@@ -37,12 +57,25 @@ class Blog extends AbstractController
      * Валидация и отправка сообщения
      * @throws \Core\RedirectException
      */
-    public function sendAction(): void
+    public function sendAction(bool $print = true)
     {
-        $imageExtension = explode('/', $_FILES['image']['type'])[1];
-        Normalizer::normalizaSpecialChars($_POST);
+        if(!$print)
+        {
+            ob_start();
+        }
+
+
+        $imageExtension = explode('\\', $_FILES['image']['type'])[1];
+        Normalizer::normalizeSpecialChars($_POST);
+
+
         if (empty($_POST['text'])) {
             $this->errors[] = EMPTY_MESSAGE_TEXT;
+        }
+
+        if(!isset($_SESSION['id']))
+        {
+            $this->errors[] = USER_NOT_AUTHORIZED;
         }
 
         if (isset($imageExtension) && !empty($errors = Validator::validateImageType($imageExtension))) {
@@ -50,34 +83,52 @@ class Blog extends AbstractController
         }
 
         if (!empty($errors = Validator::validateMessageText($_POST['text']))) {
-            $this->errors[] = $errors;
+            $this->errors = $errors;
         }
 
         if (!empty($this->errors)) {
-            echo $this->getView()->render('Blog/sendMessageErr.phtml', ['errCodes' => [$this->errors]]);
-            die;
+
+            echo $this->getView()->render($this->pathToPatterns,
+                ['errCodes' => $this->errors], 'sendErr.twig', 0, 0, 0);
+
+            return;
+
         }
 
+
         Message::send($_SESSION['id'], $_POST['text'], $imageExtension);
-        move_uploaded_file($_FILES['image']['tmp_name'], "img" . DIRECTORY_SEPARATOR . Message::getLastId() . "." . $imageExtension);
-        $this->redirect('/html/blog');
+
+        if(!isset($_FILES['image']['type'])) {
+            move_uploaded_file($_FILES['image']['tmp_name'], "img" . DIRECTORY_SEPARATOR . Message::getLastId() . "." . $imageExtension);
+        }
+
+        return $this->redirect('/html/blog');
     }
 
 
     /**
      * Вывод всех сообщений на страницу блога
      */
-    public function getAllAction(): void
+    public function getAllAction(bool $print = true): void
     {
+        if(!$print)
+        {
+            ob_start();
+        }
 
         if (!isset($_SESSION['id'])) {
             $this->errors[] = USER_NOT_AUTHORIZED;
-            echo $this->getView()->render('Blog/blogErr.phtml', ['errCodes' => $this->errors]);
-            die;
+            echo $this->getView()->render($this->pathToPatterns,
+                ['errCodes' => $this->errors], 'blogErr.twig', 0, 0, 0);
+            return;
         }
 
+
         $messages = Message::getAll();
-        echo $this->getView()->render('Blog/Blog.phtml', ['errCodes' => [$this->errors], 'messages' => $messages]);
+        echo $this->getView()->render($this->pathToPatterns
+            , ['messages' => $messages, 'pathToImages' => self::PATH_TO_IMG, 'userId' => $_SESSION['id'],
+                'admins' => ADMINS], 'Blog.twig', 0, 0, 0);
+
     }
 
 
@@ -87,7 +138,12 @@ class Blog extends AbstractController
      */
     public function deleteAction(): void
     {
-        Message::delete($_POST['id']);
-        $this->redirect('/html/blog');
+        if (in_array($_SESSION['id'], ADMINS)) {
+            Message::delete($_POST['id']);
+            $this->redirect('/html/blog');
+        }
+
+        $this->errors[] = NOT_ADMIN_ACCESS;
     }
+
 }
